fixed cart view deletes data from cart and sends infromation to a txt page

+++ OrderCookDeliver/Models/Calendar.cs
@@ -8,7 +8,7 @@
 
         public string description { get; set; }
         public string preparationTime { get; set; } 
-        public double pricePerServg { get; set; }
+        public decimal pricePerServg { get; set; }
         public string ingredient_1 { get; set; }
         public string ingredient_2 { get; set; }
         public string ingredient_3 { get; set; }
